legacy/v1/backend__old_20260105_120455/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
sqlalchemy>=2.0,<2.1
pydantic>=2.5,<3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
python-multipart==0.0.6
psycopg2-binary==2.9.9